refactor: use size_t in container.c

+++ src/container.c
@@ -40,7 +40,7 @@ void map_put(Map* map, char const* key, void* val)
 void* map_get(Map* map, char const* key)
 {
     // 逆順に探すことで新しいキーを優先する.
-    for (int i = map->keys->len - 1; i >= 0; i--)
+    for (size_t i = map->keys->len - 1; i >= 0; i--)
         if (strcmp(map->keys->data[i], key) == 0)
             return map->vals->data[i];
     return NULL;
